flaskapp/requirements.txt
Flask==2.2.2
numpy==1.21.6
openai==0.27.2
pandas==1.3.5
Werkzeug==2.2.2
matplotlib
plotly
scipy
scikit-learn 
python-dotenv
gunicorn
uWSGI==2.0.18
flask_swagger_ui
